Futurize check-gmp-none-compiler-error
This test currently emits a line number in GMP.chpl that will change if the
module is changed. Futurize the test and add a smaller reproducer. Note that
we're still tracking that we're emitting an error message with the .bad file,
but we ignore line number differences in .bad file comparisons.