Convert ZMQ to initializers
I'd failed to note that ZMQ uses the ExplicitRefCount packages, along
with the Futures module, so when converting Futures to initializers, I
also converted ExplicitRefCount, but failed to convert ZMQ.  This
resulted in internal errors due to having a mix of initializers and
constructors within a class hierarchy.

This PR converts ZMQ to initializers.  These cases seemed fairly
straightforward.