Adjust program_constructs.chpl expected LLVM
Earlier changes made the @_construct_A_chpl match the
procedure declaration, instead of the call. Then, my recent
PR caused new variables to be marked const-value-will-not-change,
so we saw more llvm invariant start annotations.

This commit simply adjusts the FileCheck directives looking
for @_construct_A_chpl to look instead for a call to a function
with that name.