Add a user-facing error message when the return type of a function is generic
Prior to this change, we would encounter some weird internal errors when the
user defined a function that returned a generic type: in normal compilation we
would get the message "every return type must also have a ref type", but generic
types don't get refTypes; when compiling with --verify we would get the message
that the type was no longer in the tree, because types that aren't used get
cleaned up and being a return type does not count as a use (only instances
created count, by causing the resolution of their type constructors).

This change adds a user-facing error message when resolving the return type of a
function.  If the type is still generic after all operations on it have
finished, then generate this error message.