Update test/classes/initializers/generics/typeFunctions/fnReturnsGeneric.chpl
This test was tracking the behavior when a type function returns a generic type.
Since we might eventually want to support this case, leave the test as a future,
but update its .bad file to track the new error message and update the .future
file with a description of the current state.  Since the user-facing error
message prevents both the normal output and output when compiling with --verify,
remove the .compopts file.