Make the error allow compilation to continue through resolution.
We'll only run into trouble after unused types have been cleaned up at the end
of resolution, and we'll encounter a USR_STOP between when the error is
discovered and then.  The first class function case shows that there are other
errors that should be generated for the same mistake, so this will allow them
to get generated as well.

Unfortunately, this points out that my error message is not good enough - in
this situation, it could lead the user to think that they can't have their
function return a generic argument - in normal cases outside of first class
functions, it will be okay because the argument will get an instantiation and
so the returned value will as well.