Modify error message and case where it gets output
Only error when in a type function, and indicate more clearly that this only
applies to type variables, not instances that are declared generic (but get
instantiated in practice).

Update the .bad file for fnReturnsGeneric.  This also means that I don't have
to update the expected output for test/functions/firstClassFns/captureGenericFn,
which is good because the error message interspersal was a bit weird.