Accept comm count increase for cylic alloc/alloc_all tests
PR #8073 removed workaround-looking code from createICField
that made the iterator capture _this arguments by value.
This workaround had already been disabled for arrays and
seems unprincipled. However, disabling the workaround caused
the `this` argument for a range iterator to be passed by reference
instead of by value.

Adjusting the default intent for ranges to be `const in` seems to
solve the issue, as would reinstating the workaround.

This commit accepts the comm counts increases so that we can
have testing pass while a more principled solution (possibly
including rangese are `const in` by default) is developed.