Limit the amount of oversubscription we do for darwin testing
On darwin we run up to 4 concurrent testing configurations (3 of which are long
running.) Historically, this oversubscription hasn't been a problem, but some
new software periodically bogs down the machine and is causing timeouts. Try to
alleviate these timeouts by limiting the number of chapel executables that can
run at once. Note that I expect this to increase total testing time. If the
increase is too much, we can limit the amount of gnu and maybe
gasnet-everything testing that we do.