Include on functions in markTaskFunctionsInIterators
markTaskFunctionsInIterators marks task functions with
FLAG_TASK_FN_FROM_ITERATOR_FN but this did not previously
include on statements. The compiler's sense of a task function
does include on statements (see that isTaskFun returns true
for on statements) and so it is reasonable to apply the same
treatment to on statements.

Resolves numa failures with
 test/parallel/taskPar/vass/yield-arrays-in-task-funs

after PR #8073.