Add more tests to lock in behavior
Ensures that the implementation would handle using a global in the if
expression, both for the method case and the function case, and that the
implementation would handle using a field in the if expression in the method
case.  The other two tests only handled referencing other arguments.

Also, fix a typo in the original tests where I was accidentally printing out
an extra line.