Convert Buffers from constructors to initializers
My confidence in this rewrite is low... particularly with respect
to the copy initializers.  It looks to me as though maybe only two
tests in the test system exercise this code, so that isn't helping
with my confidence either (and makes me wonder whether this should
be a package module rather than a standard module).