Adjust RVF serialize/deserialize for new auto destroy placement
PR #8073 migrated an auto-destroy for coforall index variables
to the task function. If the task function is also an on, there
might be RVF and serialization occurring. For the following two
tests, that was causing a yielded domain to be destroyed twice.

  studies/hpcc/STREAMS/bradc/stream-block1dpar-arr
  studies/hpcc/STREAMS/bradc/stream-block1dpar

To fix the issue, this commit adjusts insertSerialization to
explicitly handle the case in which an argument is destroyed
in the task function. If the argument is still passed in some
form, it will still be destroyed there. If the argument is no
longer passed (which is the case when there isn't a runtime type),
instead destroy the index variable immediately after the
serialize call.

The adjustment adds asserts to check that the auto-destroy
call in question is only present for coforall index variables,
since this code should be double-checked if that ever changes.