Add triple-quoted multiline uninterpreted strings
Add a new style of strings that uses either ''' or """ as the delimiter.
These are uninterpreted and allow literal newlines.  Escape sequences such
as "\n" are treated as two separate characters, not a newline.

Add a test that demonstrates the behavior of these strings.

TODO: regenerate flex/bison generated files, update documentation