Use just a single transaction for each memory region table broadcast.
In the case where we needed to broadcast both one or more memory region
table entries and a new entry count we've been doing two transactions,
one for the entries and one for the count.  Here, combine the two into a
single broadcast.  This will make some upcoming planned changes simpler,
and might marginally improve performance.  Limited 16-node testing on a
Cray XC system seems to show a slight performance benefit, but that
could be illusory.  In any event, it's hard to imagine it hurting.

While here I simplified the part of the broadcast function interface
that specifies which entries to send.  It used to take a pointer to the
first local entry to send and the number of bytes to send.  Now it takes
the index of the first entry to send and the number of entries.  The old
interface was intended to avoid any need for multiplies in the broadcast
function but the cost of those is completely overshadowed by the cost of
the network transaction, so it's better to go for simplicity and clarity
here.

I also moved the broadcast helper function.  It was in the midst of the
alloc/postAlloc/free triplet of public functions, which felt rather like
an interruption when looking at the code.  Now it follows them.