Fix issue with isAbsPath on crays
cray wasn't included in the list of Unix-like environments for isAbsPath.
This just expands the check to include it and netbsd (since the other
operatings systems we support for CHPL_TARGET_PLATFORM are really old,
I didn't feel the need to include them).  Since the check got so big,
I moved it to a helper function.

Note that in order to avoid the compiler error, I was unable to use
startsWith to condense these checks.