Update isAbsPath test so that it will fail on platforms we don't support
Previously the isAbsPath test just checked fixed string literals. Update the
test to check if Python's abspath for the CWD is detected as an absolute path.
This will let us know if our abspath implementation is wrong on other platforms
that we may test on in the future. e.g. this test would (correctly) fail on
windows since we don't support windows style paths yet.