Convert an internal error about param formal default value into two user errors
A param formal with a non-param default value or a type-mismatched default
value resulted in an internal error, as reported in issue #8240. Change the
internal error into two user errors:

error: default value for param formal is not a param
or
error: type mismatch between declared formal type and default value type

Added tests to trigger each of these errors.