Remove compiler error is isAbsPath for cygwin
I hadn't realized that cygwin's path style was POSIX instead of Windows, which
means that the current implementation of isAbsPath would work for it and that
we don't have a CHPL_TARGET_PLATFORM setting where the function would *not*
work.  Remove the compiler error branch, since we don't have a way to detect
we're on a filesystem that this wouldn't work for.