Improve hasCopyConstructor
PR #5642 added hasCopyConstructor to detect the presence of a
copy-initializer. However the apprach used in that PR (namely to loop
over all of the FnSymbol) doesn't work in the event that the type in
question is a generic type and the specific copy initializer hasn't been
resolved yet.

To resolve that issue (and at the same time to remove the loop over all
FnSymbols), change hasCopyConstructor to try resolving an appropriate
copy init call.

While there, renamed hasCopyConstructor to hasCopyInit since it's
detecting a copy initializer (vs an old-style copy constructor).

This PR is intended to resolve an issue that came up when
converting SharedObject to use initializers (see PR #8170).