Enable getInsertPointForTypeFunction to work with initializers
This function looks for the instantiation point to find where the
function such as auto-copy/auto-destroy should be added.

This logic originated in 3788ee34fa9f42bdce19e9e3cf46ccfbb1c60ac2 and I
believe it would be better to store such functions within the
instantiated type, somehow.

Anyway, it was finding the instantiation point by looking at
at->defaultInitializer->instantiationPoint. But, types with initializers
don't have a defaultInitializer set. So, this commit extends it to look
for the instantiation point from the defaultTypeConstructor.

Additionally, any AggregateType that doesn't have a defaultInitializer or
a defaultTypeConstructor simply uses the defPoint of the type as the
place to add the new functions. That makes sense since the only
complicated thing here is handling generic instantiations.