Fix memory leak introduced in PR #8073
PR #8073 enabled the ReturnByRef transformation in callDestructors to
apply to iterators. However, that created a relatively large memory
leak in the test

  test/parallel/forall/vass/reduce-expr-with-forall-in-par-iter.chpl

The AST pattern created for the forall loops in this test involves
'move'ing from a global variable to a chpl__iterLF variable. Then
newly-enabled ReturnByRef transformation sees this pattern and adds an
autoCopy.

The chpl__iterLF variable is apparently not destroyed -
FLAG_INSERT_AUTO_DESTROY does not apparently get set on it since it is
created explicitly by the compiler. Additionally, PR #5554 had already
attempted to disable the auto-copy from ReturnByRef for a related pattern.

This commit:
 * updates the move-from-global-variable case in ReturnByRef to
   respect FLAG_NO_COPY (the other cases already do)
 * updates the creation of AST with FLAG_CHPL__ITER (aka the
   chpl__iterLF variable) to always also set FLAG_NO_COPY.
 * removes the check in ReturnByRef for FLAG_CHPL__ITER since
   FLAG_NO_COPY now covers it.

In the future, I hope that our Forall AST representation will allow
us to avoid creating chpl__iterLF until after callDestructors. That
would prevent us from needing FLAG_NO_COPY in this case.