Document next idea for solving this bug and move on
Michael suggested replacing the current handling of if expressions with a new
AST node that could handle them in place.

However, it seems like the general agreement (and I think Michael is included in
this to some extent) is that this is a lower priority fix.  So document the
idea to fix it and move on for now.