The gmake clock-skew-warning filters should use a more portable regex
PR #8151 added filters to silently remove "clock skew detected" warning messages
from gmake (if any), because those messages (when they occur) caused false test
failures- both in Chapel's test harness, and in the shell script used to
implement "make check". When the modified "make check" was run on omnios, the
grep command failed because it did not accept "-E".

This change introduces a grep/regex that solves the omnios compatibility problem.

In the "make check" script (checkChplInstall) only, it also reverts some unnecessary
complexity in PR #8151.  To understand this PR, it may be helpful to
    git diff 7926fcce -- util/test/checkChplInstall
(ie, diff the version BEFORE PR #8151)