Catch more use-before-init phase 1 errors
Before this PR users could call methods on fields during phase 1, and the
compiler would fail to complain. For example:

```chpl
// 'x' is a field
proc init() {
  x[1] = 1;
  x.foo = 1;
  x.m();
  super.init();
}
```

The three statements preceeding the 'super.init' should be flagged as
'use before init' errors at compile-time. The second statement was in
fact identified as an error, but the message was confusing (can't pass
"this" as an actual to a function...).

Three changes are made to InitNormalize::fieldUsedBeforeInitialized:

1) "x[1] = 1": If the LHS of an assignment looks like a field access and the
   call is flagged with the 'square' tag, then we are dealing with a field use.

2) "x.foo = 1": If the LHS of an assignment is not a simple field access, then
   we need to walk through it to determine if it contains field uses.

3) "x.m()": When examining an arbitrary CallExpr that is not an assignment
   or field access, we need to examine its baseExpr as well as its actuals.
   In this case, the baseExpr is a field access like "this.x".

These cases now result in an error message like:
  Field used before it is initialized