Fix LLVM backend field access type and supertype TBAA handling (#8293)
Fix LLVM backend field access type and supertype TBAA handling (#8293)

[reviewed by @mppf ]

In the translation to LLVM, we were sometimes passing an incorrect Chapel type to the low-level version of the codegenLoadLLVM() and codegenStoreLLVM() functions. Nothing was wrong with the LLVM type, only the Chapel type, which is why it only became apparent while I was working on TBAA. This change fixes that problem. As a result, it is no longer necessary to carry around a fieldTbaaTypeDescriptor in a GenRet.

While there, I noticed that I was generating TBAA information the same for superclasses as for other classes, which meant that they were erroneously marked as pointers in the TBAA metadata. This change fixes that.

Expected test results have been updated accordingly.