Retire initAtomicsViaWrite future
Thanks to BHarsh's PR #8292, I believe this future can now be retired.
Though the .good file suggested that maybe this should "just work",
with more reflection, I don't think that's the case and believe that
if we want to initialize atomics in phase 1, we should do so by
adding support for initialize-time assignment of atomics by their
base type (if not normal assignments as well).  The .future here
suggests that even when filing it, I wasn't convinced it should work
(just shouldn't internal error), so in any case I think the original
problem is now resolved.