Avoid --baseline "leaks" for parallel/forall/vass/memleaks*
In #8287 I removed a "serial do" around the foralls, but rather confusingly
this seemed to result in a leak under --baseline. It turns out that we aren't
actually leaking, it's just that as of #6942 we put `freeIterator()` calls
inside a defer statement, so we were checking the memory used before
`freeIterator()` was called.

Just wrap the foralls in a new block to ensure `freeIterator()` is called
before we check the memory leaks. Note that this was only a problem under
--baseline. If the iterators are inlined the `freeIterator()` call is removed.