Use up to 120 comm domains on Aries-based systems.
We have long limited the number of uGNI comm domains we use to 30.  This
number was chosen for the original comm-ugni implementation, and was
appropriate at that time since the Gemini chip only has 32 NIC tiles.
We didn't update it when Aries arrived, although that chip has 128 NIC
tiles.  For a long time this was not a problem, since we didn't have
compute nodes with enough processors to use more than 30 comm domains.
(Generally speaking one only needs as many comm domains as one has CPUs
to initiate transactions on them.)  But with the advent of 36-core
compute nodes in Aries-based systems that's no longer true.

Here, up the limit on uGNI comm domains to 120 for Aries-based systems.
In order to make this work with the remote fork request indexes sent via
the remote CQ mechanisms, I also switched to using the 32-bit remote
event data for those encoded indexes instead of the 24-bit local event
data.  (We actually should have been doing this all along.)  And while
here, I recoded the event data (AKA instance ID) encoding and decoding
macros to make things a bit more clear.

This proves ra-atomic performance on BW36 XC nodes by 20-25%.