Changes reflecting review feedback.
Clarify how remote instance IDs are built and what they're for.

The local instance ID effectively only needs to be unique within the
node, so just use the comm domain index directly and don't bother with a
fancy encoding.

Restore the locale ID field of the remote instance ID to 18 bits.  It
fits fine in the 32-bit remote instance ID and isn't otherwise used.
Not that we'll ever be running Chapel programs on GNI-based systems that
big!

Define a single variable for the max comm domain count, set its value in
a NIC-sensitive way, and simplify the check against that max.