Revert portion of #8293 relating to field access type (#8312)
Revert portion of #8293 relating to field access type (#8312)

[merging now because master is broken, requesting post-commit review from @mppf ]

Testing of #8293 was incomplete because I forgot that start_test unsets CHPL_DEVELOPER. When a Chapel program is compiled with CHPL_DEVELOPER and CHPL_LLVM_DEVELOPER set and --llvm (outside of the test harness), the LLVM verifier fails due to invalid metadata. Testing only succeeds because start_test turns off the verifier.

This change is an exact reversion of expr.cpp and genret.h from #8293 , which is the portion related to field access types. The other files, related to the handling of superclasses, are not reverted. In hindsight, I would have done better to split the two purposes into two separate PRs.