Try disabling early dead code elimination
Since pruneResolveTree is normally running deadBlockElimination but not
deadVariableElimination, many variables are left without DefExprs. This
confuses the loop-details gathering in the context of lifetime analysis,
because now there are many DefExprs that are never initialized. Certainly
it would be possible to modify lifetime analysis to detect dead blocks,
or to modify pruneResolveTree to also eliminate dead variables. This
commit simply disables the call to deadBlockElimination in resolution.
This call was added in PR #1162 and is probably no longer necessary.