Add start_test tab completion support
Use argcomplete to provide tab completion for start_test. argcomplete hooks
into python scripts and dynamically executes them to determine a list a
completion hooks. This adds a little overhead to completions because the python
interpreter has to be spun up, but it avoids having to manually create/maintain
completions files like we do for the compiler.

To enable tab completion you should just be able to source:
$CHPL_HOME/util/devel/chpl-scripts-completion.bash