Report an error for a return inside a forall.
Resolves #8251.

When I switch to representing forall loops with ForallStmts in #5295, I omitted a call to checkControlFlow(). I am filling that in here.

There are a couple of interesting pieces in addition to that.

* We DO allow continue inside a forall. Unless that continue jumps out of the forall by naming a loop outside of the forall. I adjusted checkControlFlow() for that.

* I added some break/continue checking inside forall in scopeResolve.cpp - see handleForallGoto() (invoked from resolveGotoLabels()) - without being aware of the missing call to checkControlFlow(). That checking is insensitive to jumping outside of the forall and is now superceded by checkControlFlow(). So I removed the checking component of handleForallGoto().

I beefed up the existing test of checkControlFlow(), test_control_flow.chpl.

I also removed offending pieces from the test forall-continue-good.chpl and removed forall-continue-errors.chpl altogether in favor of the beefed-up test_control_flow.chpl