remove non-init halts from Crypto module (#8329)
[reviewed by @mppf]

Discussion after the merge of #8244 prompted closer examination of the Crypto module. While it was not possible to remove halts from init() due to the present inability to throw from initializers, this PR removes all the other halts in the module by having them throw instead.

This PR also introduces a new standard error type, IllegalArgumentError.