Allow type inference from new expressions in initializers
The following pattern has failed to compile for some time:

```chpl
class C {
  var r = new R(); // some record/class

  proc init() {
    r = new R();
    super.init();
  }
}
```

This pattern of inferring type from a new expression would result in an
internal compiler error. This internal error was the result of an
intentionally unhandled case, probably inserted to allow progress
elsewhere.

The solution is to leverage the existing paths for inferring type from
CallExprs.

This commit also introduces a compiler error if a type field is
initialized from a new expression.