Fix issue with promotion and default initializers
After the iterator memory management change, a couple of tests starting failing
with default initializers with the message "proc returns an iterator that will
be immediately converted to an array", and then a compiler internal error during
codegen.

This was because the _new function when used with promotion was not getting
marked appropriately.  This commit adds the appropriate flag to the wrapper
created in the case of promotion.

With this change, the tests that had started failing now pass again.