Split range.readWriteThis
Previously, since the compiler was (mostly) treating range like
other records, readWriteThis could work because the compiler would
specialize it for reading/writing and then infer if 'this' was modified
in it or not & use that information to set the intent of this.

But now that the default intent for range is 'const in' (including the
this intent), that approach no longer works. Instead, now range
will have readThis (with explicit ref this intent) and writeThis.