Skip init methods for building virtual maps
It is not appropriate to dynamically dispatch on initializers.  In the case of
actual calls, Ben's #8259 handles what is needed.  However, we then go on to
generate a bunch of virtual maps without following the same code path, and so
we add a version of these initializers to the hierarchy.  Default initializers
detected this when a derived class overwrote the field of a parent class.

With this change, the bug with test/classes/diten/shadow_with_different_type and
default initializers is now resolved, and another test changed its failure mode.