Fix up anonymousRangeIter test
I noticed that on master it wasn't detecting the range construction
in the coforall case. Based upon discussion with Elliot, this is due
to PR #5018 and known. So this commit removes the coforall case from
this test. Additionally it updates the prediff to check for the lines
in order so we won't be surprised again.