Really remove mem warning macros includes.
There used to be a strdup() call to assign a writable copy of "sockets"
to the provider name fabric attribute.  I replaced that with a straight
assignment in order to get the initial merge of the ofi comm layer to
pass smoke testing with minimal changes.  Here, actually do this right
by allocating space, duplicating the string into it, and assigning that
to the provider name attribute.  In support of this, add a new mem desc
type for comm layer utility space.  Also, while here, rewrite the #if
cascade that checks the CHPL_COMM_SUBSTRATE setting and remove the
memory warning macro includes since those don't exist any more.