Fix compilation segfaults for type-inferred fields
The compiler could segfault if it attempted to insert initialization
implicitly for a type-inferred field. A small example:

```chpl
class C {
  var t = new R();
  proc init() {}
}
```

The compiler will try to insert something like 'this.t = new R()' within
the initializer body.

A couple of branches in ``InitNormalize::initializeFieldsBefore`` could
call functions if the exprType was NULL, but the field's type was not
dtUnknown. However, those functions relied on exprType and so we ran
into segfaults. The solution is to remove the check against dtUnknown
and to let other existing branches handle type inference.