Avoid normalizing calls within ContextCallExprs
ContextCallExprs are generated at resolution time, but default initializers were
running into a situation where we copied a ContextCallExpr into a new function
that we then normalized and tried to resolve.  Prior to this change,
normalization would replace the contents with call_tmps, but that confused
ContextCallExprs when we tried to retrieve the expected type from it.  This
change prevents the creation of those call_tmps, so the ContextCallExpr is
handled as normal.