BUILD_VERSION_FILE runs updateBuildVersion if it exists
The old command (looking for .git directory) does not work for me since I
use git worktree. In a git worktree setup, $CHPL_HOME will be a git
checkout and respond to git commands like 'git log', but it does not have
a .git directory in it. That causes failures with
test/compflags/bradc/printstuff/version.chpl (among others) because the
BUILD_VERSION file contains "0" instead of a git hash.

I believe the rule for BUILD_VERSION_FILE checked for the .git directory
to differentiate between a Chapel release and a git checkout. But a
Chapel release does not include updateBuildVersion (and if it did, we
could try to run it!).

So, this commit changes the relevant Makefile snippet to check for an
executable UPDATE_BUILD_VERSION instead of a .git directory. That solves
the problem in my git worktree configuration.