Privatize the ofi comm layer internal interface.
This renames global symbols that are private to the ofi comm layer
(i.e., those not in the public interface) so that they all start with
"chpl_comm_ofi_".  This will help prevent collisions with user symbols.

While here I also shortened the internal "commDiagnostics" symbols to
just "commDiags" because their lengths were getting a little out of
hand.

The runtime and a user program will still build but not run, but now
grepping the symbols for things containing "ofi" but not "chpl_comm_ofi"
produces only local symbols, no globals.