Add an opt-in mechanism to reduce contention from ugni's polling thread
Add an option to have the polling thread block until an active message arrives
instead of actively polling for one. Specifically, instead of doing CqGetEvent
with a sched_yield() when there's no event just do CqWaitEvent.

The current polling scheme creates a significant amount of contention for any
task running on the same core as the polling thread. This is particularly
harmful to benchmarks like ISx, where performance is limited by the slowest
task.

Using CqWaitEvent requires that we create the CQ with GNI_CQ_BLOCKING. Creating
a CQ with GNI_CQ_BLOCKING does add some extra latency/overhead between when an
event arrives and when we can find out about it. This can unfortunately
increase the latency for AMs. This only seems to impact micro-benchmarks, but
we need to do some more investigation before enabling this by default. It's
worth noting that just creating the CQ with GNI_CQ_BLOCKING is what degrades
performance, so it's not possible to do some sort of hybrid spin/wait scheme
where we spin for a while before waiting.

This can be enabled experimentally by setting CHPL_RT_COMM_UGNI_BLOCKING_CQ=y,
which can have some significant performance improvements on real benchmarks:

 - 40% improvement for block/stencil dist versions of stencil-prk
 - 15% improvement for NPB EP
 - 10% improvement for ISx
 - 10% improvement for LLNL CoMD
 - 2% perf improvement for stream-ep (and stabilizes timings)

Regressions for micro-benchmarks:

 - 2x slowdown for multi-trial reduction
 - 20% slowdown for multi-trial coforall

Both of these micro-benchmarks end up doing 10,000 or more trials, so the
effects are really compounded. For the reduction, all remote locales do an
on-stmt to locale 0 to accumulate which hurts even more. To improve that case
we might be able to CqVectorWaitEvent, which can queue multiple events.

Adding this functionality now allows us to investigate performance on master.
Longer term we probably need a better strategy for limiting the interference
from the polling thread, hopefully without increased latency for some cases.
Whatever strategy we end up with should also be applicable to other comm layers
(so far as I know GASNet doesn't have a mechanism to do blocking AM_poll())