Fix noInlineAndHtml
Changes -t flag to --html since -t was removed in PR #8357.

While there, changed this test (that used to be a .future that
only ran with valgrind) to run in any configuration so failures
will be easier to find, and added the generated log directory
to .gitignore and a new CLEANFILES in the test's directory.