Add test for init calls that look like copy-init, but can promote
Let's say we have a class over a generic type, and this class has
promotion enabled and yields that type. Let's also say that this class
has an initializer that accepts a variable of the generic type, like so:

proc init(x : int);

If we have an expression like:

var ci = new C(int);
var x  = new C(ci);

A user would probably expect this to use some kind of copy initializer.
But if the author of ``C`` doesn't have a copy initializer for ``C(int)``,
then a naive compiler might allow the new-expr to promote. The user
would then be left with an array instead of a ``C``, which could be very
confusing.

Instead, disallow promotion for copy initializers if ``this`` and the
formal are instantiated from the same root type (i.e. ``C``). The user
is presented with an 'unresolved call' error. If the user really did
want promotion, they can easily write a forall loop themselves.