Replace _promotionType field with chpl__promotionType method
The '_promotionType' field was used to indicate and implement promotion over a
class or record. This was a hacky approach likely implemented long ago to make
progress. More recently, this field complicates the implementation and
transition for the ongoing initializers work.

Ideally we would infer promotion support by resolving the yield type of the
'these' iterator. This is not easy in the current compiler architecture. A
simpler approach is to attempt to resolve a method that returns a type. If
resolvable, promotion is enabled.

Constructors and initializers now use a common function,
'resolvePromotionType', to infer support for promotion. This function assumes
that the given type is non-generic and that it does not already have an
associated promotion type. In order to support promotion over 'this' within
initializers, the compiler will attempt to infer the promotion type before the
phase 2 portion of an initializer has been resolved.

Types that use generic initializers are fully instantiated within
'resolveInitField', at which point we can safely infer the promotion type.
Types with concrete initializers will have their promotion type inferred at the
start of 'resolveFunction'. Constructors will continue to infer promotion type
at the start of 'resolveTypeConstructor'.